fix: move get_ontology_download_url.
I think it makes more sense for it to live in supported_versions.py There is no need to initialize an OntologyParser object.
